backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
httpx>=0.26.0
pydantic[email]>=2.5.0
python-dotenv>=1.0.0
aiosqlite>=0.20.0
tenacity>=8.2.0
